µManager/backend/requirements.txt
Flask==2.3.2
Flask-RESTful==0.3.9
Flask-CORS==3.0.10
mysql-connector-python==8.0.33
